+++ sbws/util/config.py
@@ -174,6 +174,8 @@ def _validate_cleanup(conf):
     ints = {
         'stale_days': {'minimum': 1, 'maximum': None},
         'rotten_days': {'minimum': 1, 'maximum': None},
+        'v3bw_files_compress_after_days': {'minimum': 1, 'maximum': None},
+        'v3bw_files_delete_after_days': {'minimum': 1, 'maximum': None},
     }
     all_valid_keys = list(ints.keys())
     errors.extend(_validate_section_keys(conf, sec, all_valid_keys, err_tmpl))