+++ CHANGELOG.md
@@ -53,6 +53,7 @@ Summary of changes:
   Requires graphviz to be installed.(GHPR#226)
 - Add metadata to setup.py, useful for source/binary distributions.
 - Add possibility to log to system log. (#26683)
+- Add option to cleanup v3bw files. (#26701)
 
 ### Fixed